requirements-prod.txt
django-storages[google]
google-cloud-secret-manager
gunicorn
sentry-sdk[django]